rename arrow doodle and select
